% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{changeBetaNames}
\alias{changeBetaNames}
\title{Change variable names}
\usage{
changeBetaNames(names, para, para_coef, S)
}
\arguments{
\item{names}{A vector of characters. Variables taken from rstan outputs.}

\item{S}{Number of participants}

\item{Q}{Number of questions}
}
\description{
Change variable names
}
