% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_other.R
\name{.colVars}
\alias{.colVars}
\title{Little function to calculate posterior variances from simulation}
\usage{
.colVars(a)
}
\arguments{
\item{a}{}
}
\description{
Little function to calculate posterior variances from simulation
}
