% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_other.R
\name{getDIC}
\alias{getDIC}
\title{Get DIC}
\usage{
getDIC(stan_fit, devName = "dev_m")
}
\arguments{
\item{stan_fit}{Output of stan}

\item{devName}{Name for deviation in the RCode, should be default 'dev_m'}
}
\description{
Get DIC
}
