% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{getFit}
\alias{getFit}
\title{Get stanfit}
\usage{
getFit(
  stan_data,
  stan_model_compiled,
  warmup = 0,
  iter = 10,
  chains = 1,
  cores = chains,
  seed = sample(1e+05, 1),
  sample_file = NULL,
  init_r = 1,
  refresh = 500,
  adapt_delta = 0.99,
  stepsize = 0.05,
  max_treedepth = 20
)
}
\arguments{
\item{stan_data}{Formmated rstan data. Output of getRstanData().}

\item{stan_model_compiled}{stan model.}

\item{warmup}{Warm-up or burn-in number. Default 0.}

\item{iter}{Iteration number. Default 10.}

\item{chains}{Chain number. Default 1.}

\item{cores}{Number of cores to be used. Default = chain number.}

\item{seed}{Seed number}

\item{sample_file}{Save model details}

\item{init_r}{Default 1.}

\item{refresh}{Default 500}

\item{adapt_delta}{Default 0.99}

\item{stepsize}{Default 0.05}

\item{max_treedepth}{Default 20}

\item{save_model_file}{Specify a name if you want to save the RStan samples and diagnostics to a csv file.}
}
\description{
Get stanfit
}
