% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{getModelDiag}
\alias{getModelDiag}
\title{Model diagnostics.}
\usage{
getModelDiag(stan_hddm, csv_name = NULL, dev_name = "dev", logName = "log_lik")
}
\arguments{
\item{stan_hddm}{model outputs obtained from getStanFit}

\item{csv_name}{if specified then outputs will be saved in a .csv file starting with this name}

\item{dev_name}{string. name for dev in rstan codes}

\item{logName}{}
}
\value{
A dataframe indicating DIC, WAIC, model running time,
maximum rhat, minimum effective sample size,
number of divergent transitions and number of max treedepth hit.
}
\description{
\code{getModelDiag} returns model diagnostics,
and save the results to a .csv file when \code{csv_name} is specified
}
