% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{getParaSummary}
\alias{getParaSummary}
\title{Parameter summary.}
\usage{
getParaSummary(stan_data, stan_hddm, csv_name = NULL)
}
\arguments{
\item{stan_hddm}{model outputs obtained from getStanFit}

\item{csv_name}{if specified then outputs will be saved in a .csv file starting with this name}
}
\description{
\code{getParaSummary} returns parameter summaries for group and individual level parameters,
including means, percentiles, and sd.
It also saves the results to a .csv file when \code{csv_name} is specified
}
