% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_other.R
\name{getRhat}
\alias{getRhat}
\title{Get rhat}
\usage{
getRhat(stan_fit, parameters)
}
\arguments{
\item{stan_fit}{Output of stan}

\item{parameters}{Parameter names to extract}
}
\description{
Get rhat
}
