% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{getRstanData}
\alias{getRstanData}
\title{Get formatted data for rstan}
\usage{
getRstanData(
  file_name = NULL,
  data = NULL,
  a_coef = "Intercept",
  t_coef = "Intercept",
  z_coef = "Intercept",
  v_coef = "Intercept"
)
}
\arguments{
\item{file_name}{Name of the .csv data file. Columns needed include subj_idx, resp, rt, and all coefficient columns.}

\item{data}{A data frame. Columns needed include subj_idx, resp, rt, and all coefficient columns.}

\item{a_coef}{Coefficients for DDM threshold. Default: Intercept.}

\item{t_coef}{Coefficients for DDM non-decision time. Default: Intercept.}

\item{z_coef}{Coefficients for DDM starting point. Default: Intercept.}

\item{v_coef}{Coefficients for DDM drift rate. Default: Intercept.}
}
\description{
Get formatted data for rstan
}
