% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model.R
\name{runModel}
\alias{runModel}
\title{Fit the hddm}
\usage{
runModel(
  file_name = NULL,
  data = NULL,
  a_coef = "Intercept",
  t_coef = "Intercept",
  z_coef = c("Intercept", "context1", "context2"),
  v_coef = c("Intercept", "highValue", "lowValue", "context1", "context2"),
  seed = sample(1e+05, 1),
  warmup = warmup,
  iter = iter,
  chains = chains,
  cores = chains,
  csv_name_para = NULL,
  csv_name_diag = NULL,
  sample_file = NULL,
  init_r = 1,
  refresh = 500,
  adapt_delta = 0.99,
  stepsize = 0.05,
  max_treedepth = 20
)
}
\arguments{
\item{file_name}{Name of the .csv data file. Columns needed include subj_idx, resp, rt, and all coefficient columns.}

\item{data}{A data frame. Columns needed include subj_idx, resp, rt, and all coefficient columns.}

\item{a_coef}{Coefficients for DDM threshold. Default: Intercept.}

\item{t_coef}{Coefficients for DDM non-decision time. Default: Intercept.}

\item{z_coef}{Coefficients for DDM starting point. Default: Intercept.}

\item{v_coef}{Coefficients for DDM drift rate. Default: Intercept.}

\item{seed}{Seed number}

\item{warmup}{Warm-up or burn-in number. Default 0.}

\item{iter}{Iteration number. Default 10.}

\item{chains}{Chain number. Default 1.}

\item{cores}{Number of cores to be used. Default = chain number.}

\item{sample_file}{Save model details}

\item{init_r}{Default 1.}

\item{refresh}{Default 500}

\item{adapt_delta}{Default 0.99}

\item{stepsize}{Default 0.05}

\item{max_treedepth}{Default 20}
}
\description{
Fit the hddm
}
