% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_other.R
\name{summarySE}
\alias{summarySE}
\title{summarySE}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{.drop}{}
}
\description{
summarySE
}
